+++ cookbooks/aws-parallelcluster-environment/resources/nfs/nfs_ubuntu22+.rb
@@ -30,13 +30,3 @@ action :setup do
   action_install_nfs4
   action_disable_start_at_boot
 end
-
-action_class do
-  def override_server_template
-    node.default['nfs']['config']['server_template'] = '/etc/nfs.conf.d/parallelcluster-nfs.conf'
-    edit_resource(:template, node['nfs']['config']['server_template']) do
-      source 'nfs/nfs-ubuntu22+.conf.erb'
-      cookbook 'aws-parallelcluster-environment'
-    end
-  end
-end